requirements.txt
tensorflow==2.18.0
flask==3.0.0
pillow==10.1.0
numpy
matplotlib
werkzeug==3.0.1
mlflow
